requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
enum34==1.1.6
idna==2.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyYAML==3.12
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
